engine/requirements.txt
# V3.1 Engine Dependencies

# ─── Core ───
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# ─── Database ───
psycopg[binary]>=3.2.0
redis>=5.2.0

# ─── Data ───
numpy>=1.26.0
polars>=1.0.0
pandas>=2.2.0
yfinance>=0.2.40

# ─── HMM / ML ───
hmmlearn>=0.3.2
scikit-learn>=1.5.0

# ─── FinBERT ───
transformers>=4.40.0
torch>=2.2.0
# CPU 설치: pip install torch --index-url https://download.pytorch.org/whl/cpu

# ─── 통계 ───
statsmodels>=0.14.0
scipy>=1.14.0

# ─── SHAP (설명가능성) ───
shap>=0.45.0

# ─── 실행 ───
alpaca-trade-api>=3.2.0
httpx>=0.27.0

# ─── 스케줄러 ───
apscheduler>=3.10.0

# ─── gRPC ───
grpcio>=1.67.0
grpcio-tools>=1.67.0
protobuf>=5.28.0

# ─── 유틸 ───
python-dotenv>=1.0.0
python-dateutil>=2.9.0
